v0.3 Release Prep (#59)
* Removes preempt versions of async_ functions from comm

* Removes optional 'from' field from comm, all containers, and all examples.

* Feature/multimap_async_visit_group (#37)

* Adds async_visit_group functionality to ygm::container::multimap

* Adds example showcasing differences between async_visit and async_visit_group for ygm::container::multimap

* Removes unused 'pmap' argument from examples/multimap_visit_group.cpp lambdas

* Adds test of ygm::container::multimap::async_visit_group

* Removes ygm::container::disjoint_set::async_make_set function

* Feature/disjoint_set_for_all (#41)

* Adds ygm::container::disjoint_set::for_all function

* Adds example using disjoint_set for connected components

* Adds test of ygm::container::disjoint_set::for_all

* Feature/disjoint_set_async_union_and_execute (#40)

* Adds initial ygm::container::disjoint_set::async_union_and_execute functionality (does not currently handle extra arguments)

* Adds example building a spanning tree using a disjoint_set

* Expands graph and prints graph edges in disjoint_set_spanning_tree.cpp

* Adds simple test of async_union_and_execute

* Feature/disjoint_set_for_all (#41)

* Adds ygm::container::disjoint_set::for_all function

* Adds example using disjoint_set for connected components

* Adds test of ygm::container::disjoint_set::for_all

* Adds extra argument support for lambdas in ygm::container::disjoint_set::async_union_and_execute

* Adds initial ygm::container::disjoint_set::async_union_and_execute functionality (does not currently handle extra arguments)

* Resolving conflict

* Expands graph and prints graph edges in disjoint_set_spanning_tree.cpp

* Resolving conflict

* Adds extra argument support for lambdas in ygm::container::disjoint_set::async_union_and_execute

* feature/boost-guards (#43)

* Added a c++17 __has_include() guard to ygm/detail/cereal_boost_json.hpp so it will load nothing if boost is not linked. Also cleaned up executable prep function namespace.

* typo fix

* Changed from a compile guard to throwing a compiler error if __has_include() does not find Boost.JSON.

* Map async_visit_if_missing and async_visit_if_missing_else_visit (#42)

* Adds ygm::container::map::async_insert_if_missing and ygm::container::map::async_insert_if_missing_else_visit functions and associated tests and examples

* Changes ygm::container::map::insert_if_missing_else_visit to use the passed value as the first argument to the visit lambda

* Feature/actions (#46)

Added Github Actions CI testing GCC 9, GCC 10, Boost 1.77, MPICH and OpenMPI.

* Feature/mpi probe (#45) (#47)

Changed to MPI_Probe based design.

* Feature/cereal_config (#44)

* Added cereal config maker to install process.

* Removed extra namespace for consistency with spack loading. Will be a problem if spack distro changes to use cereal::cereal as the library target in the future.

* Added cereal config maker to install process.

* Removed extra namespace for consistency with spack loading. Will be a problem if spack distro changes to use cereal::cereal as the library target in the future.

* Removed config maker for cereal, and instead adapted its own. Be careful if loading cereal via spack; all downstream using cereal must also have spack loaded.

* Removed redundant export for cereal. Installation is now handled entirely by cereal source CMakeLists.txt.

* Added option to make clear what JUST_INSTALL_YGM means.


* fixed the ordering of calls in config constructor so that we find installed cereal target when including ygm in a downstream package. (#49)

* Removed extra vector copy. (#48)

* Removed extra vector copy.

* Removes shrink_to_fit() after sending message buffer


* temp fix to barrier. (#50)

* Feature/smart barrier (#51)

* Refactored barrier and added pre-barrier callback support.
* Added Control Flow barrier.
* Fixed multiple compiler warnings.
* Added CI job without boost.

* Added parallel parsers for text, csv, and ndjson. (#52)

Added features to read text, CSV, and NDJSON files in parallel.

* moved meta into detail. (#53)

* Changed to json::object (#56)

* Adds missing include statements for new I/O functionality (#55)

* Includes iomanip in ygm/io/detail/csv.hpp for access to std::quoted and filesystem in ygm/io/line_parser.hpp for access to std::filesystem namespace

* Adds stdc++fs to target_link_libraries for ygm

* Add GCC8 Test (#57)

* Overhaul of Readme (#58)

* Moved make_ygm_ptr to comm. (#54)

* Added assert.hpp include to ygm_ptr.hpp. Breaks on ASSERT_RELEASE depending upon include order otherwise. (#60)

* Iss/39 (#62)

* change cereal dependency management to allow downstream dependencies to fetch it prior to ygm.

* exchanged FetchContent_MakeAvailable for FetchContent_POPULATE + add_subdirectory.

* Updated cmake formatting; now using cmake-format.

* Added commentary as to why we are not using a versioned find_package when looking for cmake.

* Fixes bug in line_parser that calls getline when first character is not a newline. Intended behavior is to call getline when previous character is not a newline. (#64)

* Feature/contributing (#65)

* First pass on CONTRIBUTING.md file. Definitely want more content on specific style for documentation, variable/classname cases, etc, but we will need to solidfy those decisions first. We'll build this out as we go.

* Added link.

* Detect and avoid recursion in async() (#66)

* Increments YGM's version number in CMakeLists.txt (#67)

+++ test/test_bag.cpp
@@ -31,6 +31,6 @@ int main(int argc, char** argv) {
     bbag.async_insert("dog");
     bbag.async_insert("apple");
     bbag.async_insert("red");
-    ASSERT_RELEASE(bbag.size() == 3 * world.size());
+    ASSERT_RELEASE(bbag.size() == 3 * (size_t)world.size());
   }
 }
